+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Better DICOM multi-level detection ([#1196](../../pull/1196))
 - Added an internal field to report populated tile levels in some sources ([#1197](../../pull/1197), [#1199](../../pull/1199))
 - Allow specifying an empty style dict ([#1200](../../pull/1200))
+- Allow rounding histogram bin edges and reducing bin counts ([#1201](../../pull/1201))
 
 ### Changes
 - Change how extensions and fallback priorities interact ([#1192](../../pull/1192))